Cast (PARAM_ARRAY|PARAM_INPUT) en NULLAudioPlugin

+++ NULLAudioPlugin/NULLAudioPlugin.cpp
@@ -106,7 +106,8 @@ const std::string NULLAudioPlugin::g_properties[] = {
 };
 
 const int NULLAudioPlugin::g_paramTypes[] = {
-PARAM_ARRAY | PARAM_INPUT
+	// TODO: arreglar este cast
+	(int)(PARAM_ARRAY | PARAM_INPUT)
 };
 
 const int * NULLAudioPlugin::getPropertiesType() const
